be/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
rapidfuzz==3.14.3
djangorestframework-simplejwt==5.3.1

torch==2.6.0

Unidecode==1.4.0
django-cors-headers==4.4.0
django-filter==24.3
mysqlclient==2.2.4

sentence-transformers==5.1.2

sentencepiece==0.2.0
transformers==4.56.2
